tests/requirements.txt
pytest==6.2
pytest-xdist==1.31.0
pytest-timeout==1.3.4
ephemeral_port_reserve==1.1.1
bip32~=2.0
pynacl==1.4
noiseprotocol==0.3.1
toml==0.10.2
